pyproject.toml
[tool.poetry]
name = "project-telegram-bot-ip"
version = "0.1.0"
description = ""
authors = ["Anton Zimin <zimkaa87@gmail.com>"]
readme = "README.md"
packages = [{include = "project_telegram_bot_ip"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.2"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
pytz = "^2023.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
flake8-import-order = "^0.18.2"
flake8-variables-names = "^0.0.5"
flake8-commas = "^2.1.0"
flake8-builtins = "^2.1.0"
flake8-bugbear = "^23.3.23"
flake8-pyproject = "^1.2.3"
mypy = "^1.2.0"
pep8-naming = "^0.13.3"
pytest = "^7.3.1"
pylint = "^2.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-vvv"
testpaths = [
    "tests",
]

[tool.flake8]
max-line-length = 120
exclude = [
    ".git",
    "__pycache__",
    "docs/source/conf.py",
    "old",
    "build",
    "dist",
    "env",
    "venv",
    ".venv",
    ".env",
    ".python-version",
]
max-complexity = 8
extend-ignore = ["C812", "I201"]
per-file-ignores = """
    my_bot.py: E501
    __init__.py: F401
"""

[tool.mypy]
ignore_missing_imports = true
